Mobile Phone Price Detection
Hey! This project is focused on predicting the price of mobile phones using some parameters provided in the dataset. Recall that phone prices depend on a wide range of parameters. So with the development of this ML model, we have been able to ascertain those parameters that have a more significant impact on phone prices. However, the ascertainment of these parameters made it possible for the development of the ML model to aid in the prediction of phone prices with any available data.